(#57) Introduce actual color translation unit
